+++ CHANGES.rst
@@ -11,6 +11,8 @@ Improvements
 ^^^^^^^^^^^^
 
 - Include keywords in contribution CSV/Excel exports (:pr:`7421`)
+- Add new permission that grants only registration creation/editing permissions
+  (:pr:`XXXX`, thanks :user:`moliholy, unconventionaldotdev`)
 
 Bugfixes
 ^^^^^^^^